requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
pyyaml==6.0.2
pypdfium2==4.30.0
Pillow==10.4.0
loguru==0.7.2
paddleocr==2.6.1
paddlepaddle==2.6.1
numpy==1.23.5
prometheus-client
